Added Rate Limit PM and Changed Timer IDs
The bot now pm's a user if they use a command too often.  It also gives
them the number of seconds until they can use that command.  The name
that is stored in the Rate Limiter was changed from just the userid to
the userid + the command name.  This changes the paradigm from a
per-user cooldown to a per-user-command cooldown.  A single user can now
use each command with its own rate limit.