Use raw events (#28)
* Lowercase roles/emoji to make them match.

They weren't matching if there was a lowercase word in the role
due to using title case.

* Update some libraries.

* Update the max function length to 100.

Honestly I'm not sure how much max function length makes sense
in javascript where having a container function for a bunch of
other functions (which is happening here) is commom.

* Switch to handling reactions using raw events.

discord.js' handling of reactions is brittle, because it requires
all the structures it passes along to already be cached. This
switches us over to using the raw messages and fetching from
and populating caches by ourselves.