Update command infrastructure in a few ways:
Change to the less opaque "commandsPerSecond" and "commandSurgeMax"
values for commands, allowing for more configurability on frequency
of different commands.

Get rid of configurability of unlimited command execution on a per
command basis. I don't _think_ that will ever come up but if we
need to we can bring it back. Also we don't need to configure that
unlimited means '0' tickets used. That seems like a universal
constant.

Also added the Avasarala command.